Replace old particle filter
