+++ module/localisation/BallLocalisation/src/BallModel.hpp
@@ -93,7 +93,7 @@ namespace module::localisation {
         inline int getRogueCount() const {
             return n_rogues;
         }
-        Eigen::Matrix<Scalar, 2, 1> getRogueRange() const {
+        [[nodiscard]] Eigen::Matrix<Scalar, 2, 1> get_rogue() const {
             return resetRange;
         }
