+++ shared/tests/utility/math/filter/TestFilters.cpp
@@ -24,7 +24,6 @@
 
 #include "VanDerPolModel.hpp"
 
-#include "utility/math/filter/BidskiParticleFilter.hpp"
 #include "utility/math/filter/ParticleFilter.hpp"
 #include "utility/math/filter/UKF.hpp"
 
@@ -183,74 +182,6 @@ TEST_CASE("Test the UKF", "[utility][math][filter][UKF]") {
     INFO("    Process Noise.....: " << process_noise.transpose());
     INFO("    Initial State.....: " << initial_state.transpose());
     INFO("    Initial Covariance: \n" << initial_covariance);
-    model_filter.model.process_noise = process_noise;
-    model_filter.set_state(initial_state, initial_covariance);
-
-    INFO("Feeding noisy measurements into the filter")
-    std::array<double, 100> innovations;
-    std::array<std::pair<utility::math::filter::UKF<double, shared::tests::VanDerPolModel>::StateVec,
-                         utility::math::filter::UKF<double, shared::tests::VanDerPolModel>::StateMat>,
-               100>
-        actual_state;
-    for (size_t time_count = 0; time_count < 100; ++time_count) {
-        model_filter.measure(measurements[time_count], measurement_noise);
-        model_filter.time(deltaT);
-        innovations[time_count]  = measurements[time_count].x() - model_filter.get().x();
-        actual_state[time_count] = std::make_pair(model_filter.get(), model_filter.getCovariance());
-    }
-
-    INFO("Calculating statistics")
-
-    double count_x1                  = 0.0;
-    double count_x2                  = 0.0;
-    double mean_innovations          = 0.0;
-    double mean_x1_boundary          = 0.0;
-    double mean_x2_boundary          = 0.0;
-    Eigen::Vector2d mean_state_error = Eigen::Vector2d::Zero();
-    for (size_t i = 0; i < actual_state.size(); ++i) {
-        Eigen::Vector2d state_error = true_state[i] - actual_state[i].first;
-        double covariance_bounds_x1 = std::sqrt(actual_state[i].second(0, 0));
-        double covariance_bounds_x2 = std::sqrt(actual_state[i].second(1, 1));
-        mean_x1_boundary += covariance_bounds_x1;
-        mean_x2_boundary += covariance_bounds_x2;
-
-        count_x1 += (std::abs(state_error.x()) - covariance_bounds_x1) > 0.0 ? 1.0 : 0.0;
-        count_x2 += (std::abs(state_error.y()) - covariance_bounds_x2) > 0.0 ? 1.0 : 0.0;
-
-        mean_innovations += innovations[i];
-        mean_state_error += state_error;
-    }
-
-    mean_innovations /= innovations.size();
-    mean_state_error /= actual_state.size();
-    mean_x1_boundary /= actual_state.size();
-    mean_x2_boundary /= actual_state.size();
-
-    double percentage_x1 = 100.0 * count_x1 / actual_state.size();
-    double percentage_x2 = 100.0 * count_x2 / actual_state.size();
-
-    INFO("The mean of the innovations is: " << mean_innovations << ". This should be small.");
-    INFO("The mean of the state errors is: " << mean_state_error.transpose() << ". This should be small.");
-    INFO(percentage_x1 << "% of state 1 estimates exceed the 1\u03C3 boundary");
-    INFO(percentage_x2 << "% of state 2 estimates exceed the 1\u03C3 boundary");
-    INFO("The mean 1\u03C3 boundary for state 1 is [" << -mean_x1_boundary << ", " << mean_x1_boundary << "]");
-    INFO("The mean 1\u03C3 boundary for state 2 is [" << -mean_x2_boundary << ", " << mean_x2_boundary << "]");
-
-    REQUIRE(percentage_x1 <= 30.0);
-}
-
-
-TEST_CASE("Test the BidskiParticleFilter", "[utility][math][filter][BidskiParticleFilter]") {
-
-    utility::math::filter::BidskiParticleFilter<double, shared::tests::VanDerPolModel> model_filter;
-
-    INFO("Configuring the ParticleFilter with")
-    INFO("    Time step..........: " << deltaT);
-    INFO("    Number of Particles: " << number_of_particles);
-    INFO("    Process Noise......: " << process_noise.transpose());
-    INFO("    Initial State......: " << initial_state.transpose());
-    INFO("    Initial Covariance.: \n" << initial_covariance);
-    model_filter.model.n_particles   = number_of_particles;
     model_filter.model.process_noise = process_noise;
     model_filter.set_state(initial_state, initial_covariance);
