+++ shared/utility/math/filter/ParticleFilter.hpp
@@ -26,8 +26,44 @@
 #include <random>
 #include <vector>
 
+#include "utility/math/stats/multivariate.hpp"
+#include "utility/math/stats/resample/multinomial.hpp"
+#include "utility/math/stats/resample/resample.hpp"
+#include "utility/math/stats/resample/residual.hpp"
+#include "utility/math/stats/resample/stratified.hpp"
+#include "utility/math/stats/resample/systematic.hpp"
+
 namespace utility::math::filter {
 
+    using utility::math::stats::MultivariateNormal;
+
+    // The possible resampling methods
+    // See http://users.isy.liu.se/rt/schon/Publications/HolSG2006.pdf for an explanation on each one
+    enum class ResampleMethod { MULTINOMIAL, RESIDUAL, STRATIFIED, SYSTEMATIC };
+
+    /**
+     * Particle Filter
+     *
+     * @author Alex Biddulph, Trent Houliston
+     *
+     * @details Uses a set of particles (or samples) to represent possible states in a system. Each particle is randomly
+     * selected from a multivariate normal distribution. Each particle is weighted by how likely it is to be the real
+     * state of the system. Particles with a higher weight are more likely to be resampled (with noise) during the time
+     * update.
+     *
+     * Time Update: Resamples all particles, perturbing the resampled particles with the model's process noise, and asks
+     * the model to predict what should happen to each particle after an elapsed amount of time. Also known as the
+     * predict step.
+     *
+     * Measurement Update: Asks the model to predict what the next measurement will be (based on current state/particle)
+     * and the finds the difference between the model's prediction and the actual measurement. The likelihood of this
+     * measurement occuring is then determined using a multivariate normal distribution probability density function.
+     *
+     * @tparam Scalar The scalar type to use for all calculations. This puts an upper bound on the numerical precision
+     * that is achievable by the filter
+     * @tparam FilterModel The model to use in the filtering process. This model needs to provide the equations that
+     * models the system.
+     */
     template <typename Scalar, template <typename> class FilterModel>
     class ParticleFilter {
     public:
@@ -38,242 +74,396 @@ namespace utility::math::filter {
         using StateVec = Eigen::Matrix<Scalar, Model::size, 1>;
         using StateMat = Eigen::Matrix<Scalar, Model::size, Model::size>;
 
+        /**
+         * @brief Constructs the particle filter using default values. Mean is set to 0 and covariance is set to
+         * \f$0.1 * I_{n}\f$, where n is the size of the model. The residual method is used for resampling particles and
+         * the residual will use the systematic method for resampling the residual particles.
+         */
+        ParticleFilter() {
+            set_state(StateVec::Zero(), StateMat::Identity() * 0.1);
+            resample_method = ResampleMethod::RESIDUAL;
+            residual_method = ResampleMethod::SYSTEMATIC;
+        }
+
+        /**
+         * @brief Constructs the particle filter with the provided values. The resampling method with default to the
+         * residual method with the systematic method for resampling the residual particles. It is an error to use the
+         * residual method for resampling residuals and an exception will be thrown.
+         *
+         * @param mean The mean of the multivariate normal distribution. Used for sampling the initial particles in the
+         * filter.
+         * @param covariance The covariance of the multivariate normal distribution. Used for sampling the initial
+         * particles in the filter.
+         * @param resample_method_ The method to use when resampling particles during the time update step.
+         * @param residual_method_ When using ResampleMethod::RESIDUAL for the resample method, this method will be used
+         * when the residual method needs to resample residual particles.
+         */
+        ParticleFilter(const StateVec& mean,
+                       const StateMat& covariance,
+                       const ResampleMethod& resample_method_ = ResampleMethod::RESIDUAL,
+                       const ResampleMethod& residual_method_ = ResampleMethod::SYSTEMATIC) {
+            set_state(mean, covariance);
+            resample_method = residual_method_;
+
+            // The residual resampling method requires a secondary method to resample the residual particles. This
+            // secondary method cannot be the residual method.
+            if (residual_method == ResampleMethod::RESIDUAL) {
+                throw std::runtime_error(
+                    "Invalid setting for residual method. Can't use the residual resampling method to sample "
+                    "residuals.");
+            }
+            residual_method = resample_method_;
+        }
+
+        /**
+         * @brief Constructs the particle filter with the provided values. The resampling method with default to the
+         * residual method with the systematic method for resampling the residual particles. It is an error to use the
+         * residual method for resampling residuals and an exception will be thrown.
+         *
+         * This version initialises the particles using multiple initial hypotheses. The particle count is split evenly
+         * across all hypotheses. If the division between the number of hypotheses and the number of particles is not
+         * even, then the remaining particles will be treated as rogues. It is an error for the numbers of means and
+         * covariances to differ.
+         *
+         * @param mean The mean of the multivariate normal distribution. Used for sampling the initial particles in the
+         * filter. There is one mean for each initial hypothesis.
+         * @param covariance The covariance of the multivariate normal distribution. Used for sampling the initial
+         * particles in the filter.  There is one covariance for each initial hypothesis.
+         * @param resample_method_ The method to use when resampling particles during the time update step.
+         * @param residual_method_ When using ResampleMethod::RESIDUAL for the resample method, this method will be used
+         * when the residual method needs to resample residual particles.
+         */
+        ParticleFilter(const std::vector<StateVec>& mean,
+                       const std::vector<StateMat>& covariance,
+                       const ResampleMethod& resample_method_ = ResampleMethod::RESIDUAL,
+                       const ResampleMethod& residual_method_ = ResampleMethod::SYSTEMATIC) {
+            set_state(mean, covariance);
+            resample_method = residual_method_;
+
+            // The residual resampling method requires a secondary method to resample the residual particles. This
+            // secondary method cannot be the residual method.
+            if (residual_method == ResampleMethod::RESIDUAL) {
+                throw std::runtime_error(
+                    "Invalid setting for residual method. Can't use the residual resampling method to sample "
+                    "residuals.");
+            }
+            residual_method = resample_method_;
+        }
+
+        /**
+         * @brief Clears all particles in the filter and samples new particles from the multivariate normal
+         * distribution with the given mean and covariance. Sampled particles are limited (using model::limit)
+         * function to ensure all particles are valid. Particle weights are initialised to 1.
+         *
+         * @param mean The mean of the multivariate normal distribution.
+         * @param covariance The covariance of the multivariate normal distribution.
+         */
+        void set_state(const StateVec& mean, const StateMat& covariance) {
+            // Initialise all of the particles
+            init(mean, covariance);
+
+            // Limit the state of each particle to ensure they are still valid
+            for (unsigned int i = 0; i < particles.cols(); i++) {
+                particles.col(i) = model.limit(particles.col(i));
+            }
+        }
+
+        /**
+         * @brief Clears all particles in the filter and samples new particles from the multivariate normal
+         * distribution with the given mean and covariance. Sampled particles are limited (using model::limit)
+         * function to ensure all particles are valid. Particle weights are initialised to 1.
+         *
+         * This version initialises the particles using multiple initial hypotheses. The particle count is split evenly
+         * across all hypotheses. If the division between the number of hypotheses and the number of particles is not
+         * even, then the remaining particles will be treated as rogues. It is an error for the numbers of means and
+         * covariances to differ.
+         *
+         * @param mean The mean of the multivariate normal distribution. There is one mean for each initial hypothesis.
+         * @param covariance The covariance of the multivariate normal distribution. There is one covariance for each
+         * initial hypothesis.
+         */
+        void set_state(const std::vector<StateVec>& mean, const std::vector<StateMat>& covariance) {
+            // Make sure we have sane inputs
+            if (mean.size() != covariance.size() || mean.size() == 0) {
+                throw std::runtime_error("ParticleFilter::set_state called with invalid data");
+            }
+
+            // Initialise all of the particles
+            if (mean.size() == 1) {
+                init(mean[0], covariance[0]);
+            }
+            else {
+                init(mean, covariance);
+            }
+
+            // Limit the state of each particle to ensure they are still valid
+            for (unsigned int i = 0; i < particles.cols(); i++) {
+                particles.col(i) = model.limit(particles.col(i));
+            }
+        }
+
     private:
-        // Our random number generator
-        std::mt19937 rng;
-        std::normal_distribution<Scalar> norm;
+        using ParticleList    = Eigen::Matrix<Scalar, Model::size, Eigen::Dynamic>;
+        using ParticleWeights = std::vector<Scalar>;
 
-        // Dimension types for vectors and square matricies
-        using ParticleList = Eigen::Matrix<Scalar, Model::size, Eigen::Dynamic>;
+        ParticleList particles;
+        ParticleWeights weights;
+        ResampleMethod resample_method;
+        ResampleMethod residual_method;
 
-        ParticleList sample_particles(const StateVec& mean, const StateMat& covariance, const int& n_particles) {
-            // Sample single gaussian (represented by a gaussian mixture model of size 1)
-
-            // Implementation based on the work presented in
-            // Conrad Sanderson and Ryan Curtin.
-            // An Open Source C++ Implementation of Multi-Threaded Gaussian Mixture Models, k-Means and
-            // Expectation Maximisation. International Conference on Signal Processing and Communication
-            // Systems, 2017.
-
-            ParticleList new_particles =
-                ParticleList::NullaryExpr(Model::size, n_particles, [&]() { return norm(rng); });
-
-            const StateMat sqrt_covariance = covariance.cwiseSqrt();
-
-            for (int i = 0; i < n_particles; ++i) {
-                new_particles.col(i) = sqrt_covariance * new_particles.col(i) + mean;
+        /**
+         * @brief Clears all particles in the filter and samples new particles from the multivariate normal
+         * distribution with the given mean and covariance. Particle weights are initialised to 1.
+         *
+         * @param mean The mean of the multivariate normal distribution.
+         * @param covariance The covariance of the multivariate normal distribution.
+         */
+        void init(const StateVec& mean, const StateMat& covariance) {
+            // Make sure our particle list has the right shape
+            if (particles.cols() != model.n_particles + model.n_rogues || particles.rows() != Model::size) {
+                particles.resize(Model::size, model.n_particles + model.n_rogues);
+                weights.resize(model.n_particles + model.n_rogues);
             }
 
-            return new_particles;
+            // Setup our multivariate normal distribution so we can initialise the particles
+            MultivariateNormal<Scalar, Model::size> multivariate(mean, covariance);
+
+            // Sample a random vector from the multivariate distribution for each particle
+            for (int i = 0; i < model.n_particles; ++i) {
+                particles.col(i) = multivariate.sample();
+            }
+
+            // Get the model to give us some rogues
+            for (int i = 0; i < model.n_rogues; ++i) {
+                particles.col(model.n_particles + i) = model.get_rogue();
+            }
+
+            // Start all weights at 1
+            std::fill(weights.begin(), weights.end(), Scalar(1));
         }
 
-        ParticleList sample_particles(const StateVec& mean, const StateVec& covariance, const int& n_particles) {
-            // Sample single gaussian (represented by a gaussian mixture model of size 1)
-
-            // Implementation based on the work presented in
-            // Conrad Sanderson and Ryan Curtin.
-            // An Open Source C++ Implementation of Multi-Threaded Gaussian Mixture Models, k-Means and
-            // Expectation Maximisation. International Conference on Signal Processing and Communication
-            // Systems, 2017.
-
-            ParticleList new_particles =
-                ParticleList::NullaryExpr(Model::size, n_particles, [&]() { return norm(rng); });
-
-            const StateVec sqrt_covariance = covariance.cwiseSqrt();
-
-            for (int i = 0; i < n_particles; ++i) {
-                new_particles.col(i) = new_particles.col(i).cwiseProduct(sqrt_covariance) + mean;
+        /**
+         * @brief Clears all particles in the filter and samples new particles from the multivariate normal
+         * distribution with the given mean and covariance. Particle weights are initialised to 1.
+         *
+         * This version initialises the particles using multiple initial hypotheses. The particle count is split evenly
+         * across all hypotheses. If the division between the number of hypotheses and the number of particles is not
+         * even, then the remaining particles will be treated as rogues.
+         *
+         * @param mean The mean of the multivariate normal distribution.
+         * @param covariance The covariance of the multivariate normal distribution.
+         */
+        void init(const std::vector<StateVec>& mean, const std::vector<StateMat>& covariance) {
+            // Make sure our particle list has the right shape
+            if (particles.cols() != model.n_particles + model.n_rogues || particles.rows() != Model::size) {
+                particles.resize(Model::size, model.n_particles + model.n_rogues);
+                weights.resize(model.n_particles + model.n_rogues);
             }
-
-            return new_particles;
-        }
-
-        [[nodiscard]] ParticleList sample_particles(const std::vector<StateVec>& mean,
-                                                    const std::vector<StateMat>& covariance,
-                                                    const int& n_particles) {
-            const size_t mean_size = mean.size();
-            if (mean_size != covariance.size() || mean_size == 0) {
-                throw std::runtime_error("ParticleFilter::set_state called with invalid data");
-            }
-
-            if (mean_size == 1) {
-                return sample_particles(mean[0], covariance[0], n_particles);
-            }
-
-            // Sample single gaussian (represented by a gaussian mixture model of size 1)
-
-            // Implementation based on the work presented in
-            // Conrad Sanderson and Ryan Curtin.
-            // An Open Source C++ Implementation of Multi-Threaded Gaussian Mixture Models, k-Means and
-            // Expectation Maximisation. International Conference on Signal Processing and Communication
-            // Systems, 2017.
-
-            ParticleList new_particles =
-                ParticleList::NullaryExpr(Model::size, n_particles, [&]() { return norm(rng); });
 
             // We want to evenly distribute the particles across all hypotheses
-            // If this is not an even division, the last N particles will just be randomly initialised
-            const int particles_per_state = n_particles / mean_size;
+            // If the distribution is not even then the remaining particles will be treated as extra rogues
+            const int particles_per_state = model.n_particles / mean.size();
 
-            for (int state = 0; state < int(mean_size); ++state) {
-                const StateMat sqrt_covariance = covariance[state].cwiseSqrt();
+            // For each hypothesis sample a random vector from the multivariate distribution for each particle
+            for (int hypothesis = 0; hypothesis < int(mean.size()); ++hypothesis) {
+                // Setup our multivariate normal distribution for this hypothesis so we can initialise the particles
+                MultivariateNormal<Scalar, Model::size> multivariate(mean[hypothesis], covariance[hypothesis]);
 
-                const int start_col = state * particles_per_state;
+                const int start_col = hypothesis * particles_per_state;
                 for (int i = 0; i < particles_per_state; ++i) {
-                    new_particles.col(start_col + i) = sqrt_covariance * new_particles.col(start_col + i) + mean[state];
+                    particles.col(start_col + i) = multivariate.sample();
                 }
             }
 
-            return new_particles;
+            // Get the model to give us some rogues
+            const int num_particles = particles_per_state * mean.size();
+            const int num_rogues    = model.n_particles - num_particles + model.n_rogues;
+            for (int i = 0; i < num_rogues; ++i) {
+                particles.col(num_particles + i) = model.get_rogue();
+            }
+
+            // Start all weights at 1
+            std::fill(weights.begin(), weights.end(), Scalar(1));
         }
 
-        [[nodiscard]] ParticleList sample_particles(const std::vector<StateVec>& mean,
-                                                    const std::vector<StateVec>& covariance,
-                                                    const int& n_particles) {
-            const size_t mean_size = mean.size();
-            if (mean_size != covariance.size() || mean_size == 0) {
-                throw std::runtime_error("ParticleFilter::set_state called with invalid data");
+        /**
+         * @brief Use the chosen resample and residual method to resample the particles in the filter based on their
+         * weights.
+         */
+        std::vector<int> resample_particles() {
+
+            // Select the resampling method to use
+            switch (resample_method) {
+                case ResampleMethod::MULTINOMIAL:
+                    return stats::resample::multinomial(model.n_particles, weights.begin(), weights.end());
+                case ResampleMethod::RESIDUAL:
+                    // The residual method uses a secondary method when it resamples the residual particles
+                    switch (residual_method) {
+                        case ResampleMethod::MULTINOMIAL:
+                            return stats::resample::residual(model.n_particles,
+                                                             weights.begin(),
+                                                             weights.end(),
+                                                             stats::resample::multinomial<decltype(weights.begin())>);
+                        case ResampleMethod::STRATIFIED:
+                            return stats::resample::residual(model.n_particles,
+                                                             weights.begin(),
+                                                             weights.end(),
+                                                             stats::resample::stratified<decltype(weights.begin())>);
+                        case ResampleMethod::SYSTEMATIC:
+                            return stats::resample::residual(model.n_particles,
+                                                             weights.begin(),
+                                                             weights.end(),
+                                                             stats::resample::systematic<decltype(weights.begin())>);
+                        default: throw std::runtime_error("Invalid setting for residual method."); break;
+                    }
+                    break;
+                case ResampleMethod::STRATIFIED:
+                    return stats::resample::stratified(model.n_particles, weights.begin(), weights.end());
+                case ResampleMethod::SYSTEMATIC:
+                    return stats::resample::systematic(model.n_particles, weights.begin(), weights.end());
+                default: throw std::runtime_error("Invalid setting for resample method."); break;
+            }
+        }
+
+        /**
+         * @brief Do a weighted resampling of all of the particles in the filter. Resampled particles are then perturbed
+         * (using model::noise) and the model is asked to provide new rogue particles (if the model is using
+         * them). Resampled particles are then limited (using model::limit) to ensure validity and all particle weights
+         * are reset to 1.
+         *
+         * @param dt The amount of time that has elapsed since the previous time update step.
+         */
+        void resample(const Scalar& dt) {
+
+            // Make a multivariate normal distribution with zero mean and the model's process noise as the covariance
+            // This will be used to perturb the resampled particles
+            MultivariateNormal<Scalar, Model::size> multivariate(model.noise(dt));
+
+            // Get indexes to the particles that are being resampled
+            // Some particles may be resampled multiple times
+            std::vector<int> idx = resample_particles();
+
+            // Resample the particles
+            // For each index get the corresponding particle and perturb it with the model's process noise
+            ParticleList resampled_particles(Model::size, particles.cols());
+            for (int i = 0; i < int(idx.size()); ++i) {
+                // Create a new particle by using the selected particle as the mean + value from process noise
+                resampled_particles.col(i) = particles.col(idx[i]) + multivariate.sample();
             }
 
-            if (mean_size == 1) {
-                return sample_particles(mean[0], covariance[0], n_particles);
+            // Get the model to give us some rogue particles
+            for (int i = 0; i < model.n_rogues; ++i) {
+                resampled_particles.col(model.n_particles + i) = model.get_rogue();
             }
 
-            // Sample single gaussian (represented by a gaussian mixture model of size 1)
-
-            // Implementation based on the work presented in
-            // Conrad Sanderson and Ryan Curtin.
-            // An Open Source C++ Implementation of Multi-Threaded Gaussian Mixture Models, k-Means and
-            // Expectation Maximisation. International Conference on Signal Processing and Communication
-            // Systems, 2017.
-
-            ParticleList new_particles =
-                ParticleList::NullaryExpr(Model::size, n_particles, [&]() { return norm(rng); });
-
-            // We want to evenly distribute the particles across all hypotheses
-            // If this is not an even division, the last N particles will just be randomly initialised
-            const int particles_per_state = n_particles / mean_size;
-
-            for (int state = 0; state < int(mean_size); ++state) {
-                const StateVec sqrt_covariance = covariance[state].cwiseSqrt();
-
-                const int start_col = state * particles_per_state;
-                for (int i = 0; i < particles_per_state; ++i) {
-                    new_particles.col(start_col + i) =
-                        new_particles.col(start_col + i).cwiseProduct(sqrt_covariance) + mean[state];
-                }
+            // Update our particles with our resampled particles and limit each one to ensure it remains valid
+            for (int i = 0; i < particles.cols(); i++) {
+                particles.col(i) = model.limit(resampled_particles.col(i));
             }
 
-            return new_particles;
+            // Reset all weights to 1
+            std::fill(weights.begin(), weights.end(), Scalar(1));
         }
 
     public:
-        ParticleFilter(const StateVec& initial_mean      = StateVec::Zero(),
-                       const StateMat& initialCovariance = StateMat::Identity() * 0.1)
-            : rng(), norm() {
-            set_state(initial_mean, initialCovariance);
-        }
-        ParticleFilter(const std::vector<StateVec>& initial_mean, const std::vector<StateMat>& initialCovariance)
-            : rng(), norm() {
-            set_state(initial_mean, initialCovariance);
-        }
-
-        void set_state(const StateVec& initial_mean, const StateMat& initialCovariance) {
-            particles =
-                sample_particles(initial_mean, initialCovariance, model.getParticleCount() + model.getRogueCount());
-
-            // Limit the state of each particle to ensure they are still valid
-            for (unsigned int i = 0; i < particles.cols(); i++) {
-                particles.col(i) = model.limit(particles.col(i));
-            }
-        }
-
-        void set_state(const std::vector<StateVec>& initial_mean, const std::vector<StateMat>& initialCovariance) {
-            particles =
-                sample_particles(initial_mean, initialCovariance, model.getParticleCount() + model.getRogueCount());
-
-            // Limit the state of each particle to ensure they are still valid
-            for (unsigned int i = 0; i < particles.cols(); i++) {
-                particles.col(i) = model.limit(particles.col(i));
-            }
-        }
-
-        void set_state(const std::vector<StateVec>& initial_mean, const std::vector<StateVec>& initialCovariance) {
-            particles =
-                sample_particles(initial_mean, initialCovariance, model.getParticleCount() + model.getRogueCount());
-
-            // Limit the state of each particle to ensure they are still valid
-            for (unsigned int i = 0; i < particles.cols(); i++) {
-                particles.col(i) = model.limit(particles.col(i));
-            }
-        }
-
+        /**
+         * @brief Resample all particles in the filter and set their weights back to 1. The model is asked to give a
+         * prediction on what happens to each particle after dt time has elapsed (using model::time).
+         *
+         * @tparam Args the types of any extra arguments that are provided to the model's time update function.
+         *
+         * @param dt The amount of time that has elapsed since the previous time update step.
+         * @param params Any extra arguments that are to be forwarded to the model's time update function.
+         */
         template <typename... Args>
-        void time(const Scalar& dt, const Args&... params) {
-            // Time update our particles
-            for (unsigned int i = 0; i < particles.cols(); ++i) {
-                particles.col(i) = model.time(particles.col(i), dt, params...);
-            }
+        void time(const Scalar& dt, Args&&... params) {
 
-            // Perturb them by noise
-            particles += sample_particles(StateVec::Zero(), StateVec(model.noise(dt).diagonal()), particles.cols());
+            // Resample our particles
+            resample(dt);
 
-            // Limit the state of each particle to ensure they are still valid
-            for (unsigned int i = 0; i < particles.cols(); i++) {
-                particles.col(i) = model.limit(particles.col(i));
+            // Get the model to apply a time update to each particle
+            for (int i = 0; i < particles.cols(); ++i) {
+                particles.col(i) = model.time(particles.col(i), dt, std::forward<Args>(params)...);
             }
         }
 
+        /**
+         * @brief Get the model to predict (using model::predict) what the next measurement will be then determine how
+         * accurate this prediction is (using model::difference). The likelihood of the measurement is determined using
+         * the multivariate normal distribution probability density function, this likelihood (determined for each
+         * particle) re-weights each particle.
+         *
+         * @tparam MeasurementScalar The underlying numerical type of the measurement
+         * @tparam S The dimensionality of the measurement (and associated covariance matrix)
+         * @tparam VArgs The types of any extra arguments associated with constructing the measurement vector.
+         * @tparam MArgs The types of any extra arguments associated with constructing the measurement covariance
+         * matrix.
+         * @tparam Args the types of any extra arguments that are provided to the model's predict function.
+         *
+         * @param measurement The measurement (from sensors, etc).
+         * @param measurement_covariance The covariance of the measurement.
+         * @param params Any extra arguments that are to be forwarded to the model's predict function.
+         */
         template <typename MeasurementScalar, int S, int... VArgs, int... MArgs, typename... Args>
         MeasurementScalar measure(const Eigen::Matrix<MeasurementScalar, S, 1, VArgs...>& measurement,
-                                  const Eigen::Matrix<MeasurementScalar, S, S, MArgs...>& measurement_variance,
-                                  const Args&... params) {
+                                  const Eigen::Matrix<MeasurementScalar, S, S, MArgs...>& measurement_covariance,
+                                  Args&&... params) {
 
-            ParticleList candidate_particles               = ParticleList::Zero(Model::size, particles.cols());
-            candidate_particles.leftCols(particles.cols()) = particles;
-
-            // Resample some rogues
-            for (int i = 0; i < model.getRogueCount(); ++i) {
-                candidate_particles.col(i + model.getParticleCount()) =
-                    particles.col(i) + model.getRogueRange().cwiseProduct(StateVec::Random() * 0.5);
+            // Get the model to make a prediction for each particle
+            // Also get the model to tell us how much a prediction deviates from the observed measurement
+            Eigen::Matrix<MeasurementScalar, S, Eigen::Dynamic> differences(S, particles.cols());
+            for (int i = 0; i < particles.cols(); ++i) {
+                differences.col(i) =
+                    model.difference(model.predict(particles.col(i), std::forward<Args>(params)...), measurement);
             }
 
-            Eigen::Matrix<MeasurementScalar, S, Eigen::Dynamic> differences(S, particles.cols());
-            for (int i = 0; i < candidate_particles.cols(); ++i) {
-                differences.col(i) = model.difference(model.predict(particles.col(i), params...), measurement);
+            // Recalculate the weights of each particle
+            // The weight update factor comes from the density function of a zero mean, measurement_covariance
+            // covariance multivariate normal distribution
+            MultivariateNormal<Scalar, S> multivariate(measurement_covariance);
+            for (int i = 0; i < particles.cols(); ++i) {
+                weights[i] *= multivariate.density(differences.col(i));
             }
 
             // Calculate log probabilities
             Eigen::Matrix<MeasurementScalar, Eigen::Dynamic, 1> logits =
-                -differences.cwiseProduct(measurement_variance.inverse() * differences).colwise().sum();
+                -differences.cwiseProduct(measurement_covariance.inverse() * differences).colwise().sum();
 
-            // Subtract the max log prob for numerical stability and then exponentiate
+            // Subtract the max log probability for numerical stability and then exponentiate
             logits = Eigen::exp(logits.array() - logits.maxCoeff());
 
-            // Resample
-            std::discrete_distribution<> multinomial(logits.data(), logits.data() + logits.size());
-            for (unsigned int i = 0; i < particles.cols(); i++) {
-                particles.col(i) = model.limit(candidate_particles.col(multinomial(rng)));
-            }
+            // Return the mean log probability
             return logits.mean();
         }
 
-        // Take the mean of all particles, pay no attention to the type of the data
-        StateVec get() const {
+        /**
+         * @brief Calculates and returns the mean of all particles.
+         */
+        [[nodiscard]] StateVec get() const {
             return particles.rowwise().mean();
         }
 
-        StateMat getCovariance() const {
+        /**
+         * @brief Calculates and returns the covariance of the particles.
+         */
+        [[nodiscard]] StateMat getCovariance() const {
             auto mean_centered = particles.transpose().rowwise() - particles.transpose().colwise().mean();
             return (mean_centered.transpose() * mean_centered) / (particles.transpose().rows() - 1);
         }
 
-        const ParticleList& getParticles() const {
+        /**
+         * @brief Returns the underlying particle list.
+         */
+        [[nodiscard]] const ParticleList& getParticles() const {
             return particles;
         }
 
-    private:
-        ParticleList particles;
+        /**
+         * @brief Returns the underlying particle weights.
+         */
+        [[nodiscard]] const ParticleWeights& getParticleWeights() const {
+            return weights;
+        }
     };
 }  // namespace utility::math::filter
